integrate StringUtils method to CapString
